feat: add copy button and inline Rescue Key actions
